Give users more options for calculating a packed message type from some non-artio input, add a message type getter on the header

+++ artio-codecs/src/main/java/uk/co/real_logic/artio/util/AsciiBuffer.java
@@ -65,7 +65,7 @@ public interface AsciiBuffer extends DirectBuffer
      */
     String getAscii(int offset, int length);
 
-    int getMessageType(int offset, int length);
+    long getMessageType(int offset, int length);
 
     DecimalFloat getFloat(DecimalFloat number, int offset, int length);
